Feature #5541: Support "read" database connection load balancing.
Now we can have multiple DSNs configured for the 'db' connection. A 'db' DSN will be selected randomly on weights. It means that we can set the chance for a connection to be used, e.g. A connection weight 2 will have twice the chance to be used than a connection weight 1.
If Matrix fails to connect to a 'read' DB in the DSN pool, it will log the error and try to use another DSN. It will throw "Fail to connect to any DB in the DSN Pool" exception when there is no DSN left in the Pool. However, to keep backward compatibility, it will throw the same error like the previous version if there is just one DSN configured for 'db' connection.