Bug #5696 Errors with uploading allowed file types when 'Allowed File Extensions to Upload' settings have been customised.
The JS API was creating the file extension based on the supplied Matrix file type, which means it would assume a single file extension per file type. For example a video file would always be created as 'wmv', regardless of the supplied filename. But since the file would be renamed eventually to the original filename (via subsequent EES requests), it wasn't creating any noticiable issues.

However if file extension restrictions (Matrix global prefs) are imposed then JS API might fail to create the file. For example, say global file preferences are set to allow 'mp4' file extesnion only and we request JS_API to create a file 'xyz.mp4' of 'Video_File' type file asset, then the JS_API will try to create the Video_File asset with filename 'xyz.mp4.wmv', which would fail the Matrix file pref check, since 'wmv' is not listed in global file prefs as allowed type.

After this fix, the JS_API will create the file asset based on extension in the supplied filename and the given file type. So if the supplied filename's extension is one of the allowed extenions of given file type then it will create the filename with same extension as in the supplied filename. This will ensure the Matrix global file prefs are applied correctly when JS_API creates the file asset.