Bug #6166: If a field that is part of the primary key, has been altered, we will have different set of 'old' and 'new' primary keys for a rollback table. For such case, we will need to insert a new rollback entry rather then updating the existing rollback entry (as currently being done, hence the issue) that has same 'eff_from' timestamp as the session timestamp.
Changing the content type in the bodycopy div creates such scenrio, where the current content type asset is morphed to the parent 'content_type' asset and the resulting asset to the 'target' content type asset. These operations involves altering the 'attrid' (part of the table's primary key)  value in the sq_ast_attr_val table. Refer to the bug report of the details.