Bug #6298: Currently when doing global keyword replacement for the global keywords globals_get, globals_post, globals_cookie and globals_session, Matrix strip outs all the characters in the value expect for limited 'safe' characters ('A-Za-z0-9\.\-\_\%$\#, :\/'), for security reasons.
However its it not very practical to whitelist of safe characters since we can't possibly built the list of all the safe chars out there (like Polish utf8 chars as reported in the bug report, which is currently getting stripped off). Instead what we do now is escape the special html chars in the value, i.e. allow all the chars except the blacklisted ones (special html chars), which are encoded.

All the special html chars are encoded in POST and GET global vars, expect in SESSION and COOKIE where quotes are not encoded