Combining the binoculars code from the asset tag line and the Bodycopy DIV's "Multi-linked DIV" lineages popup, and splitting it out into its own function, get_asset_tag_line_binoculars().
Code re-use reasons first off, but also ensures code that uses the new function is compatible with the new asset map. There was no point in maintaining two versions of the comms between a lineages popup and the asset map.
The new function also supports a custom title to allow the Multi-Linked DIV version to keep its own title.

Side effects: the tag line version of the popup now receives a genericised version of the message that appeared in the Multi-Linked DIV popup. Used to say "This DIV is linked to 5 assets", now it's "This asset is linked to 5 parents".