Backport to 4.18: Reverting reload_assets for the asset map to use strings rather than arrays.
Tried to get this working for cleaner code, but it's not going to work because it's
trying to use an array in a different frame. Because array class definitions are part
of the window, trying to use an array in another frame will result it being treated
as a generic object, which caused problems.

So reverting to the old method of passing a string of pipe-separated asset IDs -
at least strings have always been generic types.