Fix to #6654, re drag-and-drop in IE8.
IE8 has problems with elementFromPoint() - when trying to get the element underneath
a draggable, it returns null the first time, but the correct element the second time (usually).
This seems to be because in IE, elementFromPoint() targets have to respond to mouse events. My hypothesis: once we hide the draggable to get the next-to-top element, IE8's CSS handling is slow enough to not provide mouse events yet to that element, so it returns null.
We don't want it to be truly unlimited for obvious reasons. I'll give it 4 attempts in case it's extra-slow, but IE8 will generally get it on the second attempt.