Fixed bug #6661: Restore Root in Modern Asset Map doesn't allow return to Root (#1) Folder when a custom root is set.
Apparently the custom root set in preferences is NOT a "chroot jail" - it was coded in the Java asset map to return to the custom root if it's not showing (eg. teleported elsewhere), but to switch to the root folder (#1) if it's already showing the custom root, allowing access to the full tree.