Fixed a bug that saw the bottom ~75 pixels and rightmost ~30 pixels of the asset map as potentially unusable for move-me (including "create here") clicks.
When determining whether a click was outside the scrollbars, we compared
clientWidth/Height of the tree vs the event clientX/Y. But this is
comparing apples + oranges: the clientX/Y is from the frame viewport's
view, which is not adjusted for where the tree actually starts.

Therefore these clicks were seen to be "on or beyond the scrollbars".
This caused propagation to be stopped before the move-me mode mousedown
could occur.

Squizmap #4703; Matrix bug tracker #6876.