Fix #5313: Line breaks in <pre> tags are not preserved
There was code in the Matrix Embed Movie plugin's PHP side that removed
extra whitespace as well as its original remit (to replace movies with
"Embedded Object" images).
It's been in there since the dawn of time (literally 2003) and if it was
needed for any older browser support (talking IE5) then it's not needed
now, especially as it squashes PRE tags and anything else using
white-space: pre CSS.

Also needed changes in Matrix Embed YouTube because Embed Movie was used
as the template for that.