Fix #5261: Remove SQ_ACTION=login param after you log in.
This has also been done in this commit with SQ_ACTION=logout for
cleanliness, but is more important for SQ_ACTION=login because that page
would otherwise be the result of a POST, with the obvious repercussions
from hitting "reload" in the browser.

The redirect preserves other query strings but not hashes, as the latter
are not passed to PHP.

Redirect in login is done at a similar point to precedent with
removing "_login" suffix. In logout it's done after all the
required items of logging out are done.