Fix #5393: Locating asset where sort order has gaps.
The "get assets" command now has a setting of "sortorder" which can be used instead of "start" - this will hit the database for a list of sort orders for TYPE_1/2 links underneath an asset, find the real position of the sortorder (aligning to the set size specified by the limit param or the default setting if not supplied), then using that as the "start" param.

This is intended to be used primarily for locateAsset(), where we get
passed a bunch of sort-orders direct from the database, but you may have
gaps which may be large enough for the targeted asset to not appear on
that page at all (eg. a sort order of 3000 but it actually appears as
2500th in the list), or that sort-order number not corresponding to a
page at all (eg. a sort order of 3000 when there are only 2500 assets).

Also, made start/sortorder and limit optional. If neither of start or sortorder is supplied, the default start point will be zero (or set 1). If limit is omitted, the default asset set size (in global config) will be used.

The return for each parent asset will contain a "real_order" attribute,
if sortorder is supplied, which is what it should use for drawing the
tree.

It doesn't work with bridges yet (since they generally don't put links into sq_ast_lnk - it either manages it themselves and/or dump into shadow links table). I'm not sure how feasible that will be but it'll default to using the sortorder attribute as "start" (suitably aligned to a set boundary) if the sort order can't be found.