First attempt at fix #5393: Locating asset where sort order has gaps.
The "get assets" command now has a setting of "sortorder" which can be used instead of "start" - this will hit the database for a list of sort orders for TYPE_1/2 links underneath an asset, find the real position of the sortorder (aligning to the set size specified by the limit param or the default setting if not supplied), then using that as the "start" param.

Also, made start/sortorder and limit optional. If neither of start or sortorder is supplied, the default start point will be zero (or set 1). If limit is omitted, the default asset set size (in global config) will be used.

It doesn't work with bridges yet (since they generally don't put links into sq_ast_lnk - it either manages it themselves and/or dump into shadow links table). I'm not sure how feasible that will be but it'll default to using the sortorder attribute as "start" (suitably aligned to a set boundary) if the sort order can't be found.

TODO: need to test it on Whiskey tomorrow, and then put into stable (for 5.0.2.0) when given all-clear.