Get the #5393 fix cleaned up now I can see it in action.
Mainly this was in tweaking the JS file to use it, and working out the
best place to pass the real_order attribute back to the JS.

Also cleaned up parse errors and removed error_logs/console debugs.
Also remove post-localisation change DB error msg from previous commit,
which is silly to have in 4.18.