Cleaning up the "auto add remaps" button in Dev (5.1) only.
For some reason, instead of using the backend outputter default
"read_access='write:lookups'" setting in the edit interface file, the
auto add remaps option created its own whil

This bothers me because it reduces predictability for getting
information from that screen for use by Edit+. The first field is used
for the Web Paths section, the second field for the Current URLs, but to
get the auto-remap field it is currently the FOURTH field, not the third.

This change is NOT going into stable, as it is purely cosmetic to users, and changing it
would mean Edit+ 5.0 would have to reference it, which would break the
automatic remaps feature (#4807) in older 5.0 versions. Edit+ will be
aligned with Matrix in this for 5.1.