Linking between bridge and property level complete.
Also changed the caching to have a "stale but useful" level in cases
where we don't get new data beyond the 4-hour mark. To do this, the
actual details are stored in the cache manager indefinitely, and a
refresh entry gets the 4-hour expiry.
TODO: Property-to-view linking.
TODO: What to happen if Cache Manager is disabled??