Management API-esque summary of active Content Experiments.
You won't get more than this until much later on, but given I'm working
with pulling in goals for the summary, may as well go the whole hog and
drag in experiments too.

Some changes to View (besides the report) to allow the selection of
common site- and page-level metrics (and potentially dimensions although
that's less valuable) through a constant. This is the equivalent of the
"get site report data" and "get page report data" functions in Squiz
Analytics' toolset.