Tech report for GA.
Currently a bit of a long page because it contains eight reports:
- Operating system stats for Desktop, Mobile, Tablet;
- Browser stats for Desktop, Mobile, Tablet;
- Device stats for Mobile and Tablet.

There are a lot of massaging rules to the above which include the
following:

- OSes:
... Big-cat names for OS X (at version level);
... Android is grouped at OS level by the "sweet things" codename (eg. "Android Jelly Bean" for versions 4.1-4.3) with sub-versions beneath.
... iOS is similarly grouped at OS level by major version (eg. "iOS 7")
with sub-versions beneath.
... At the moment, keeping the Linux 32-bit/64-bit versions that come
through in the version level, with 32-bit (i386-i686), 64-bit (x86-64)
and 64-bit (Itanium) being the split-up.

- Browsers, with rapid releases grouped for brevity:
... Firefox is grouped into their ESR streams (so "17-23", "24-30", etc.). A "4-9"
exists for rapid releases pre-ESR. Before rapid releases, two version
parts are provided (eg. 3.6);
... Chrome is split into years of release, with 2013 split up into
Webkit and Blink eras;
... Opera is similarly split into years of release for its Blink years
(15+), with "Presto" versions ("classic" Opera up to version 12) grouped
together.