Bridge learned "period", View learns how to use it.
Period attribute on bridge is a selection of 1/3/6 months (for now) with
the key being the ISO8601 period string that it uses (eg. "P3M" for 3
months).

Also, OS X versioning is switched to NOT_SET when it provides an
architecture but not a version.
(TODO: combining Intel and PowerPC versions of OS X that had both -
Leopard (10.5) and Tiger (10.4) - or possibly separating them out as
eg. "Leopard (10.5) on Intel".