Implement #5565: Invariant error codes for JS API.
Wherever there is an "error" section in the JS API there will be an
error code provided. This is so things like Edit+ can rely on these
codes and not the messages, which are liable to change due to editorial
fixes and especially localisation.

Messages will be broad-based if they are a common theme (eg. permission
errors, asset not found), but can be relatively specific at times (eg.
the Tidy ones).