Expose errors in fetching, and JS API learns details in returnError().
1. GA Bridge's fetchWithBackoff() now returns a three-element array:
result, errors, and [status] code. If either of the first two doesn't
apply, they will be NULL.
2. GA View's generateReport() will now return either a numeric-indexed
array if successful, or an array with one key "errors" as failure.
3. JS API's returnError() learns third parameter, $detail. This is
intended as extra data sent from internal functions or external API
calls, such as the GA one. Errors in the "Get GA Statistics" method will
be passed to this.
This will appear in the JS API return as "errorDetail", only if passed.