Better method of passing JS API's shared secret.
It will still support key in the GET for standard, but enhanced will
(instead of the original POST or the changed GET) instead use an
Authorization header like used OAuth.

The Authorization header will be in the form:
Authorization: Matrix-JSAPI key="1234567890"

Where "1234567890" is the key in the JS API asset.