First stage of concept requested in #5493.
If the result of a message is a "success" text message, also provide
other relevant information.

TODO: haven't implemented clone's results as per #5493 yet. Returning
the asset IDs of the clones but not the names and linkids and such
because the HIPO doesn't give us that. Extra work required for this.

Also gave invariant error codes to the Form Question errors.