Fix an issue with the screen compiling of section notes.
It was possible to have multiple section notes in the generated screens
(which breaks when you try to view it). That's because existing section notes
were being checked as if it were an attribute (copy-pasta'd from
display_name checking), when it's actually a child tag.
Fix the SimpleXML reference to ensure only one section note is printed.