Replace cache manager with a file-based fetch cache.
The problem with cache manager is that it is written for Matrix
permissions and thus cannot handle a "global cache" that transends users
- without any of the "levels of caching" (public, group, permission) set, nothing is cached.

So moving it to data directory (in [connector data path]/fetch-cache).
The key will be sha1'd for a filename and there will be two levels of subfolders
in order to mitigate a "maximum number of inodes" issue.

It will by default have a refresh period of 4 hours although eventually
it will probably be managed by a cron job, even if the design of the
interface is not cached as in Squiz Analytics (ie. it will request its
data and store it without actually doing anything with it).

Also storing the raw JSON response rather than the result rows, because
the response also contains things like column headers.