Changed view caching back to storing the row data only.
The reason why we were storing only row data to begin with was to allow
queries that require multiple pages of results (more than 10,000
records) to be requested.

Only this time, actually implement the multiple page stuff. The cache
key now no longer has the index and results in there.

The query itself has the relevant dimension and metric data so it's OK.