Fix a long-standing problem with lineages in assets with a __data URL.
There was code added in 2005 to handle the situation of files with a
__data URL not having a proper root URL, because their lookups are
stored with a root_urlid of zero (commit 2e219c1 onward). There, it gets
the lookup of a parent with the correct root URL, which generally has a
full lineage because files can't generally have files as parents (due to
possibility of __data), and then generate the lineage from that.

There were two issues with this:
1/ It may have worked for a static root URL but it didn't for __data,
because the code would have found the valid root URL anyway and bypassed
the code. This meant it did not return the whole lineage, or no lineage at all.
Now specifically triggers for __data.
2/ It did not add the last part - the link between the parent being used
as a surrogate, and the file asset itself. This is now added using
existing loaded assets and the significant link between them.

Fixes Edit+ Squizmap #5605. It was manifesting in "no lineage found"
errors in getLineageByURL JS API, which was causing files with __data
not to be searchable in backend search of Edit+. It doesn't apply to
Admin Mode backend search because no checks for correct lineage are made
there.