Cherry-pick 50e3c29 into 5.0.
Fix a long-standing problem with lineages in assets with a __data URL.

Ensure __data URLs go through the alternative method of getting the
parent, finding a real lineage, and then adding ourselves to the end of
it.

Manifests itself as "no lineage found" errors, eg. when used as JS API,
which was a symptom of Edit+ Squizmap #5605.