Convert trigger_localised_error() calls.
These will now have three fixed parameters: code, message and level. The
message will simply use translate() and sprintf() (as applicable) when
called which means it will no longer be fixed to the error code.

For this, I've vacated raiseError() in Locale Manager and moved it into
trigger_localised_error(), with appropriate changes. Locale Manager's
role is now in calling translate() at the calling function level.

This raises a TODO: eventually getErrorMessage() will be vacated too,
but User asset uses this for determining the various errors for
violating the password rules. This needs to be removed and replaced with
something more suitable to the "no longer fixed to error code"
environment. (This may also make it easier to provide a custom error
message.)