Localisation work on the Preferences.
Problem is that we can't just have the localisation at a level of
painting the global preferences interface, because then that doesn't
flow into the interface for customising user preferences.

So setting these in the constructor rather than in the declaration,
because otherwise they can't have translate() called on them. (Currently as
a separate variable but I will merge them in to pref_vars.)

Also taking advantage of cleaner lambdas in PHP 5.3 by writing a
function that will allow a basic field wrap-around that will print the
(translated) name and description, and then allow definition of the guts
in a closure function.

To aid simplification of prefs files I've also built functions which
build basic text and combo box functionality with one line (plus the
lines that build the option list in the latter).