More localisation and refactoring of preferences.
Prefs now have two more pre-canned options, for Int Text Boxes, and Asset Finder (including
Multiple Asset Finders).
Prefs currently modified to show a highlight behind the settings that
currently use pre-canned options, with an intent to get it as high as
possible. Some won't be able to, but we can try :)