Fix Squizmap #6181.
Specifying extra attributes in JS API's createAsset() doesn't work for
Matrix 5.1.
This is because of somewhat dodgy handling of the "attributes" option
(by appending it to the end of extra_attributes' value, thereby making
it look like its own query string vars), which worked when posting
multipart/form-data but not now it gets sent as application/json.

Handled by allowing all attributes (with appropriate handling for
certain fields) into the query, through modifying the _options function
to include an argument as to whether to discard options not in the
default string or not (defaulting to true).

Tested using both the "additional keys" and the "attributes key" method.