JS translation file work.
This will get done at the end of step 3, after all the packages are
installed, and will use the installed package list. JS file will be
generated for each locale with a string marked "js-translate" in the flags.

Also updated the get_strings extraction script. This will currently
generate empty PO files broken into groups of 150 strings so that they
can be easily given to multiple resources to work upon.