First stage of plural translations.
This is the equivalent of GNU ngettext(), or Edit+'s _n().

Format is translate_plural($singular_str, $plural_str, $number).

By default, it will use $singular_str when $number is 1, $plural_str
when it's something else. $number can be passed NULL to represent
"positive infinity", for instance when a text box is being shown instead
of a number (eg. "[TEXT BOX] assets").

The intention is that we will use the "Plural-Forms" header of the core
PO file, which has a standard format, to determine which number form
to use given which number. These should be written so the singular form
is form 0, and the "positive infinity" version is the last form.

A couple of proofs-of-concept have been added so far:
- To the first screen of the Clone Assets HIPO job;
- To the Asset Map Width user preference;
- To the Messages Per Page inbox preference.