More translation stuff.
We now have translate() markings around display names and notes for edit
interface files. This is a rare exception to the "do not put a variable
in the translate() call" situation, because the string-getting function
instead gets them directly from the edit interface files.

The remainder of these changes fall into the following categories:
- Plurals that need to be changed to translate_plural() - eg. those that
  added an "s" when a certain variable was > 1, or added (s) to be
  ignorant to the number.
- Translate calls which needed modification because the way they were
  done didn't make sense (eg. Workflow approve/reject link reactions)
- Translate calls which were piecemeal and would better fit as more
  complete strings (eg. Created date on Settings screen).
- Strings that no attempt was made to translate them .