Fix Squizmap #6264.
In certain circumstances, Simple Edit Users were still not able to
access the simple Asset Map in classic WYSIWYG plugins in Edit Mode.
This was when the root of the current domain did not have a Site asset
attached to it (all Sites on that domain were of subdirectories).

Moving the handling for the asset map request further up. We don't do
checks for whether we are on a valid asset in Admin Mode, so it probably
doesn't make sense in Edit Mode.