New addition to JS API getGeneral().
Now returns a list of URLs through the "urls" endpoint of any general
info returns. The "webpath" key still contains the closest one.

URLs in that array are just strings with the "http://" or "https://"
added if applicable.

Satisfies Edit+ Squizmap #5949, allowing Analytics screen to select a
URL different from the editing domain.