Fix Squizmap #6286.
If (1) a teleport root had been set, (2) the view had been switched to
the actual root folder (#1) and (3) the view was then refreshed, the
request for the root folder's children was limited by "Per Set"
limitation. (Pagination was omitted correctly.)

Always request all the children if the requested parent assetid is "1",
as well as if it's the teleport root.