Fixing New Child menu in IE8.
Using Array.map is not a good idea in code that needs to be IE8
compatible (and we don't have a polyfill for it like Edit+). Just do
things the old-fashioned way.

(This was reported as Squizmap #6218, but will be deleted because it's in
unreleased code and thus not an actual bug.)